Drop down menus
Added drop down menu (spinner) for profile selection to the player 1 and
2 screens. Both access string array in strings.xml in the values folder.
So the next step is to make an array of all the profile names that were
entered and have that be the array the spinner uses. Kind of minor right
now but we should also figure out how to not show the profile that was
selected for player 1 when player 2 chooses their profile. Also removed
action bar since I don't see why we need it. Could always put it back
later though.