Drop
Drop down menu when choosing profile now works with the names the user
entered in the create profile page. Also changed one button to an
imagebutton. To change the others just go to the xml of the activity
with the button you want to change, then switch <Button to <ImageButton,
add a thing that says android:src="@drawable/image" below it with image
being the name of your picture. Put your icon pictures in the
drawable-xxhdpi folder which is under res. Then in the java class for
each activity change Button to ImageButton near the top and below under
where I wrote //Locate buttons in activity_main.xml or whatever xml you
edited